python/requirements.txt
ipdb
ipython
pyyaml
